Revert "Accidental execution of tool syntax fix (#3456)"
This reverts commit 07db3c8824919afb830007aadd5d2126d4c39c1f.
