+++ src/core/assistant-message/index.ts
@@ -1,3 +1,2 @@
 export { type AssistantMessageContent, parseAssistantMessage } from "./parseAssistantMessage"
 export { presentAssistantMessage } from "./presentAssistantMessage"
-export { parseAssistantMessageV2 } from "./parseAssistantMessageV2"
